chore: divide memoRoom service
